feat(ui): implement RecyclerView with custom pagination and ViewModel logic
- Added RecyclerView logic with a custom adapter to handle pagination manually.
- Created a scroll listener to detect when the user reaches the end of the list and trigger data loading.
- Implemented ViewModel logic to fetch and manage paginated data from the service using offset and limit.
- Enabled Data Binding in MainActivity and integrated it with the XML layout for the RecyclerView.
- Updated MainActivity to observe LiveData from the ViewModel and dynamically update the RecyclerView adapter.
- Ensured smooth appending of new items to the RecyclerView without replacing the existing list.

This commit provides a fully custom pagination solution for loading items from the service, integrated with ViewModel and Data Binding.

+++ app/src/main/java/com/example/melichallenge/data/remote/models/ItemApiModel.kt
@@ -19,7 +19,7 @@ data class ItemApiModel(
     @SerializedName("thumbnail") var thumbnail: String? = null,
     @SerializedName("currency_id") var currencyId: String? = null,
     @SerializedName("order_backend") var orderBackend: Int? = null,
-    @SerializedName("price") var price: Int? = null,
+    @SerializedName("price") var price: Double? = null,
     @SerializedName("original_price") var originalPrice: String? = null,
     @SerializedName("sale_price") var salePrice: SalePriceApiModel? = null,
     @SerializedName("available_quantity") var availableQuantity: Int? = null,
@@ -41,7 +41,7 @@ data class ItemApiModel(
 
 data class SalePriceApiModel(
     @SerializedName("price_id") var priceId: String? = null,
-    @SerializedName("amount") var amount: Int? = null,
+    @SerializedName("amount") var amount: Double? = null,
     @SerializedName("currency_id") var currencyId: String? = null,
     @SerializedName("exchange_rate") var exchangeRate: String? = null,
 )
